vision_agent_evolve/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-agent-evolve"
version = "0.1.0"
description = "Self-evolving vision agent with tool and skill generation"
authors = [{name = "Vision Agent Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0,<2",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = ["core", "tools", "skills", "evolution"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
